Move RecordTracedCommand/Handle to Command base; remove TracedCommandBufferCmd and CuDnnCmd stubs
- Move TracedCommandBuffer (caching helper), Handle (RecordAction dispatcher),
  and RecordTracedCommand from TracedCommandBufferCmd to the Command base class
  as protected methods, making the tracing infrastructure available to all
  Command subclasses without requiring the intermediate base class.
- Remove TracedCommandBufferCmd from command_buffer_cmd.h/.cc.
- Remove CuDnnCmd from command_buffer_cmd.h/.cc (will move to CuDnnThunk).
- Update GemmCmd and CublasLtCmd to inherit directly from Command and override
  IsTracedCommand() to return true.
- Remove kTracedCommand from CommandType enum (no longer used).
